feat(cli)!: remove top-level 'agents set'; nest run model/mode defaults under 'agents models' (RUSH-2579)
`agents set <selector> --model/--mode` moves to `agents models set`, matching
the surface-consolidation pattern (#2621): `models` already owns model/mode
concerns (tier overrides live there too), so the short front door for run
defaults nests under it instead of standing alone at the top level.

Behavior is unchanged -- same selector forms, same flags, same run.defaults
store read by `agents config get run.<agent@version>.*`. `agents set` is now
retired (RETIRED_TOP_LEVEL_COMMANDS), so it reports unknown command instead of
silently auto-correcting to a near neighbor.
